requirements.txt
celery
Django
djangorestframework
flower
redis
django-celery-beat

pytest
pytest-django
